fix import

+++ FunctionModule/listings/import_csv.py
@@ -779,9 +779,9 @@ def handle_import(request, file_path, listing_type):
                                     listing_fisrt.save()
                                     logger.info(f"row {line_count}: cập nhật {listing} prior {listing.priority} same listdate từ {new_listing}")
 
-                        if count_update >= 1:
+                        if count_update >= 1 and listing is not None:
                             listing.delete()
-                            print(f"row {line_count}: xóa listing thừa: {listing}")
+                            print(f"row {line_count}: xóa listing rác: {listing}")
 
                         count_update += 1
                 else:
